pennywise_backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
python-telegram-bot==20.7
pandas==2.1.3
numpy==1.25.2
matplotlib==3.8.2
plotly==5.17.0
psycopg2-binary==2.9.9
requests==2.31.0
python-decouple==3.8
django-cors-headers==4.3.1
gunicorn==21.2.0
pytest==7.4.3
pytest-django==4.7.0